requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
docstring-parser==0.15
filelock==3.12.0
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
lit==16.0.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
progressbar==2.5
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
threadpoolctl==3.1.0
timm==0.5.4
tomli==2.0.1
# torch==2.0.0
torch-geometric==2.3.1
# torch-scatter==2.1.1 -f https://pytorch-geometric.com/whl/torch-1.10.0+cu113.html
# torch-sparse==0.6.13 -f https://pytorch-geometric.com/whl/torch-1.10.0+cu113.html
# torchvision
tornado==6.3.1
tqdm==4.65.0
triton==2.0.0
typed-argument-parser==1.8.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
visdom==0.2.4
websocket-client==1.5.1
Werkzeug==2.3.3
yapf==0.33.0
